+++ tests/test_util.py
@@ -13,32 +13,27 @@ def same_list(list1, list2):
         [item in list2 for item in list1] + [item in list1 for item in list2]))
 
 
-class MockPrognostic(Prognostic):
+class PrognosticPropertiesContainer(object):
 
-    def __init__(self):
-        self._num_updates = 0
-
-    def __call__(self, state):
-        self._num_updates += 1
-        return {}, {'num_updates': self._num_updates}
+    def __init__(self, input_properties, tendency_properties, diagnostic_properties):
+        self.input_properties = input_properties
+        self.tendency_properties = tendency_properties
+        self.diagnostic_properties = diagnostic_properties
 
 
-class MockImplicit(Implicit):
+class ImplicitPropertiesContainer(object):
 
-    def __init__(self):
-        self._a = 1
+    def __init__(self, input_properties, diagnostic_properties, output_properties):
+        self.input_properties = input_properties
+        self.diagnostic_properties = diagnostic_properties
+        self.output_properties = output_properties
 
-    def __call__(self, state):
-        return self._a
 
+class DiagnosticPropertiesContainer(object):
 
-class MockDiagnostic(Diagnostic):
-
-    def __init__(self):
-        self._a = 1
-
-    def __call__(self, state):
-        return self._a
+    def __init__(self, input_properties, diagnostic_properties):
+        self.input_properties = input_properties
+        self.diagnostic_properties = diagnostic_properties
 
 
 def test_update_dict_by_adding_another_adds_shared_arrays():
@@ -87,64 +82,342 @@ def test_get_component_aliases_with_no_args():
     assert len(aliases.keys()) == 0
 
 
-def test_get_component_aliases_with_single_component_arg():
-    components = [MockPrognostic(), MockImplicit(), MockDiagnostic()]
-    for c, comp in enumerate(components):
-        aliases = get_component_aliases(comp)
-        assert type(aliases) == dict
-        if c == 3:
-            assert len(aliases.keys()) == 2
-            for k in ['T', 'P']:
-                assert k in list(aliases.values())
-        else:
-            assert len(aliases.keys()) == 0
+def test_get_component_aliases_prognostic():
+    aliases = get_component_aliases(
+        PrognosticPropertiesContainer(
+            input_properties={
+                'temperature': {
+                    'alias': 'T',
+                }
+            },
+            tendency_properties={
+                'wind': {
+                    'alias': 'u'
+                }
+            },
+            diagnostic_properties={
+                'specific_humidity': {
+                    'alias': 'q'
+                },
+            }
+        )
+    )
+    assert aliases == {'temperature': 'T', 'wind': 'u', 'specific_humidity': 'q'}
 
 
-class DummyProg1(Prognostic):
-    input_properties = {'temperature': {'alias': 'T'}}
-    tendency_properties = {'temperature': {'alias': 'TEMP'}}
-
-    def __init__(self):
-        self._a = 1
-
-    def __call__(self, state):
-        return self._a
-
-
-class DummyProg2(Prognostic):
-    input_properties = {'temperature': {'alias': 't'}}
-
-    def __init__(self):
-        self._a = 1
-
-    def __call__(self, state):
-        return self._a
-
-
-class DummyProg3(Prognostic):
-    input_properties = {'temperature': {}}
-    diagnostic_properties = {'pressure': {}}
-    tendency_properties = {'temperature': {}}
-
-    def __init__(self):
-        self._a = 1
-
-    def __call__(self, state):
-        return self._a
+def test_get_component_aliases_no_alias_prognostic():
+    aliases = get_component_aliases(
+        PrognosticPropertiesContainer(
+            input_properties={
+                'temperature': {
+                }
+            },
+            tendency_properties={
+                'wind': {
+                }
+            },
+            diagnostic_properties={
+                'specific_humidity': {
+                },
+            }
+        )
+    )
+    assert aliases == {}
 
 
-def test_get_component_aliases_with_different_values():
-    # two different aliases in the same Component:
-    aliases = get_component_aliases(DummyProg1())
-    assert len(aliases.keys()) == 1
-    assert aliases['temperature'] == 'TEMP'
-    # two different aliases in different Components:
-    aliases = get_component_aliases(DummyProg1(), DummyProg2())
-    assert len(aliases.keys()) == 1
-    assert aliases['temperature'] == 't'
-    # NO aliases in component
-    aliases = get_component_aliases(DummyProg3)
-    assert len(aliases.keys()) == 0
+def test_get_component_aliases_empty_prognostic():
+    aliases = get_component_aliases(
+        PrognosticPropertiesContainer(
+            input_properties={},
+            tendency_properties={},
+            diagnostic_properties={}
+        )
+    )
+    assert aliases == {}
+
+
+def test_get_component_aliases_diagnostic():
+    aliases = get_component_aliases(
+        DiagnosticPropertiesContainer(
+            input_properties={
+                'temperature': {
+                    'alias': 'T',
+                }
+            },
+            diagnostic_properties={
+                'specific_humidity': {
+                    'alias': 'q'
+                },
+            }
+        )
+    )
+    assert aliases == {'temperature': 'T', 'specific_humidity': 'q'}
+
+
+def test_get_component_aliases_no_alias_diagnostic():
+    aliases = get_component_aliases(
+        DiagnosticPropertiesContainer(
+            input_properties={
+                'temperature': {
+                }
+            },
+            diagnostic_properties={
+                'specific_humidity': {
+                },
+            }
+        )
+    )
+    assert aliases == {}
+
+
+def test_get_component_aliases_empty_diagnostic():
+    aliases = get_component_aliases(
+        DiagnosticPropertiesContainer(
+            input_properties={},
+            diagnostic_properties={}
+        )
+    )
+    assert aliases == {}
+
+
+def test_get_component_aliases_implicit():
+    aliases = get_component_aliases(
+        ImplicitPropertiesContainer(
+            input_properties={
+                'temperature': {
+                    'alias': 'T',
+                },
+                'input2': {
+                    'alias': 'in2',
+                }
+            },
+            output_properties={
+                'wind': {
+                    'alias': 'u'
+                }
+            },
+            diagnostic_properties={
+                'specific_humidity': {
+                    'alias': 'q'
+                },
+            }
+        )
+    )
+    assert aliases == {
+        'temperature': 'T', 'input2': 'in2', 'wind': 'u', 'specific_humidity': 'q'}
+
+
+def test_get_component_aliases_no_alias_implicit():
+    aliases = get_component_aliases(
+        ImplicitPropertiesContainer(
+            input_properties={
+                'temperature': {},
+            },
+            output_properties={
+                'wind': {}
+            },
+            diagnostic_properties={
+                'specific_humidity': {},
+            }
+        )
+    )
+    assert aliases == {}
+
+
+def test_get_component_aliases_input_over_diagnostic():
+    aliases = get_component_aliases(
+        ImplicitPropertiesContainer(
+            input_properties={
+                'temperature': {'alias': 'correct'},
+            },
+            output_properties={},
+            diagnostic_properties={
+                'temperature': {'alias': 'incorrect'},
+            }
+        )
+    )
+    assert aliases == {'temperature': 'correct'}
+
+
+def test_get_component_aliases_input_over_output():
+    aliases = get_component_aliases(
+        ImplicitPropertiesContainer(
+            input_properties={
+                'temperature': {'alias': 'correct'},
+            },
+            output_properties={
+                'temperature': {'alias': 'incorrect'},
+            },
+            diagnostic_properties={}
+        )
+    )
+    assert aliases == {'temperature': 'correct'}
+
+
+def test_get_component_aliases_output_over_diagnostic():
+    aliases = get_component_aliases(
+        ImplicitPropertiesContainer(
+            input_properties={
+            },
+            output_properties={
+                'temperature': {'alias': 'correct'},
+            },
+            diagnostic_properties={
+                'temperature': {'alias': 'not'},
+            }
+        )
+    )
+    assert aliases == {'temperature': 'correct'}
+
+
+def test_get_component_aliases_diagnostic_over_tendency():
+    aliases = get_component_aliases(
+        PrognosticPropertiesContainer(
+            input_properties={
+            },
+            tendency_properties={
+                'temperature': {'alias': 'not'},
+            },
+            diagnostic_properties={
+                'temperature': {'alias': 'correct'},
+            }
+        )
+    )
+    assert aliases == {'temperature': 'correct'}
+
+
+def test_get_component_aliases_input_over_tendency():
+    aliases = get_component_aliases(
+        PrognosticPropertiesContainer(
+            input_properties={
+                'temperature': {'alias': 'T'},
+            },
+            tendency_properties={
+                'temperature': {'alias': 'not'},
+            },
+            diagnostic_properties={
+            }
+        )
+    )
+    assert aliases == {'temperature': 'T'}
+
+
+def test_get_component_aliases_empty_implicit():
+    aliases = get_component_aliases(
+        ImplicitPropertiesContainer(
+            input_properties={},
+            output_properties={},
+            diagnostic_properties={}
+        )
+    )
+    assert aliases == {}
+
+
+def test_get_component_aliases_all_types_no_overlap():
+    aliases = get_component_aliases(
+        ImplicitPropertiesContainer(
+            input_properties={
+                'input1': {
+                    'alias': 'in1',
+                }
+            },
+            output_properties={
+                'output1': {
+                    'alias': 'out1'
+                }
+            },
+            diagnostic_properties={
+                'diagnostic1': {
+                    'alias': 'diag1'
+                },
+            }
+        ),
+        DiagnosticPropertiesContainer(
+            input_properties={
+                'input2': {
+                    'alias': 'in2',
+                }
+            },
+            diagnostic_properties={
+                'diagnostic2': {
+                    'alias': 'diag2'
+                },
+            }
+        ),
+        PrognosticPropertiesContainer(
+            input_properties={
+                'input3': {
+                    'alias': 'in3',
+                }
+            },
+            tendency_properties={
+                'tendency3': {
+                    'alias': 'tend3'
+                }
+            },
+            diagnostic_properties={
+                'diagnostic3': {
+                    'alias': 'diag3'
+                },
+            }
+        )
+    )
+    assert aliases == {
+        'input1': 'in1', 'output1': 'out1', 'diagnostic1': 'diag1',
+        'input2': 'in2', 'diagnostic2': 'diag2', 'input3': 'in3',
+        'tendency3': 'tend3', 'diagnostic3': 'diag3'}
+
+def test_get_component_aliases_all_types_with_overlap():
+    aliases = get_component_aliases(
+        ImplicitPropertiesContainer(
+            input_properties={
+                'input1': {
+                    'alias': 'in1',
+                }
+            },
+            output_properties={
+                'output1': {
+                    'alias': 'out1'
+                }
+            },
+            diagnostic_properties={
+                'diagnostic1': {
+                    'alias': 'diag1'
+                },
+            }
+        ),
+        DiagnosticPropertiesContainer(
+            input_properties={
+                'input1': {
+                    'alias': 'in1',
+                }
+            },
+            diagnostic_properties={
+                'diagnostic1': {
+                    'alias': 'diag1'
+                },
+            }
+        ),
+        PrognosticPropertiesContainer(
+            input_properties={
+                'input1': {
+                    'alias': 'in1',
+                }
+            },
+            tendency_properties={
+                'tendency1': {
+                    'alias': 'tend1',
+                }
+            },
+            diagnostic_properties={
+                'diagnostic1': {
+                    'alias': 'diag1'
+                },
+            }
+        )
+    )
+    assert aliases == {
+        'input1': 'in1', 'output1': 'out1', 'diagnostic1': 'diag1', 'tendency1': 'tend1'}
 
 
 def test_ensure_no_shared_keys_empty_dicts():